Refactoring browser sockets to its own module
