+++ lib/transport/websocket.js
@@ -15,7 +15,7 @@
 var util = require('../util.js');
 var log = require('../log.js');
 var serializer = require('../serializer.js');
-
+var BrowserWsTransport = require('./websocket-browser.js')
 
 function Factory (options) {
   var self = this;
@@ -217,80 +217,10 @@ Factory.prototype.create = function () {
       });
 
     })();
-   //
-   // running in the browser
-   //
   } else {
-
-    (function () {
-
-      var websocket;
-
-      // Chrome, MSIE, newer Firefox
-      if ("WebSocket" in global) {
-
-        if (self._options.protocols) {
-          websocket = new global.WebSocket(self._options.url, self._options.protocols);
-        } else {
-          websocket = new global.WebSocket(self._options.url);
-        }
-        websocket.binaryType = 'arraybuffer';
-
-      // older versions of Firefox prefix the WebSocket object
-      } else if ("MozWebSocket" in global) {
-
-        if (self._options.protocols) {
-          websocket = new global.MozWebSocket(self._options.url, self._options.protocols);
-        } else {
-          websocket = new global.MozWebSocket(self._options.url);
-        }
-      } else {
-        throw "browser does not support WebSocket or WebSocket in Web workers";
-      }
-
-      websocket.onmessage = function (evt) {
-        log.debug("WebSocket transport receive", evt.data);
-
-        var msg = transport.serializer.unserialize(evt.data);
-        transport.onmessage(msg);
-      }
-
-      websocket.onopen = function () {
-        var serializer_part = websocket.protocol.split('.')[2];
-        for (var index in self._options.serializers) {
-          var serializer = self._options.serializers[index];
-          if (serializer.SERIALIZER_ID == serializer_part) {
-            transport.serializer = serializer;
-            break;
-          }
-        }
-
-        transport.info.protocol = websocket.protocol;
-        transport.onopen();
-      }
-
-      websocket.onclose = function (evt) {
-        var details = {
-          code: evt.code,
-          reason: evt.message,
-          wasClean: evt.wasClean
-        }
-        transport.onclose(details);
-      }
-
-      // do NOT do the following, since that will make
-      // transport.onclose() fire twice (browsers already fire
-      // websocket.onclose() for errors also)
-      //websocket.onerror = websocket.onclose;
-      transport.send = function (msg) {
-        var payload = transport.serializer.serialize(msg);
-        log.debug("WebSocket transport send", payload);
-        websocket.send(payload);
-      }
-      transport.close = function (code, reason) {
-        websocket.close(code, reason);
-      };
-    })();
+    // running in the browser
+    this.browserWsTransport =
+      new BrowserWebsocketTransport(transport, this._options)
   }
 
   return transport;
